+++ src/Sentry.AspNetCore/SentryMiddleware.cs
@@ -8,7 +8,6 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Sentry.Extensibility;
-using Sentry.Protocol;
 using Sentry.Reflection;
 
 namespace Sentry.AspNetCore;
@@ -99,10 +98,13 @@ internal class SentryMiddleware : IMiddleware
                 // sent to Sentry. This avoid the cost on error-free requests.
                 // In case of event, all data made available through the HTTP Context at the time of the
                 // event creation will be sent to Sentry
-                scope.OnEvaluating += (_, _) =>
+
+                // Important: The scope that the event is attached to is not necessarily the same one that is active
+                // when the event fires.  Use `activeScope`, not `scope` or `hub`.
+                scope.OnEvaluating += (_, activeScope) =>
                 {
-                    SyncOptionsScope(hub);
-                    PopulateScope(context, scope);
+                    SyncOptionsScope(activeScope);
+                    PopulateScope(context, activeScope);
                 };
             });
 
@@ -167,11 +169,11 @@ internal class SentryMiddleware : IMiddleware
         }
     }
 
-    private void SyncOptionsScope(IHub newHub)
+    private void SyncOptionsScope(Scope scope)
     {
         foreach (var callback in _options.ConfigureScopeCallbacks)
         {
-            newHub.ConfigureScope(callback);
+            callback.Invoke(scope);
         }
     }